pyproject.toml
[tool.poetry]
name = "love-story-generator"
version = "0.1.0"
description = ""
authors = ["cyn <chao.yn@foxmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.1"
pyyaml = "^6.0.2"
requests = "^2.32.3"
openai = "^1.62.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
